rename bitextor_main.py to bitextor_cli.py, and add main_full function that allows to call snakemake directly
